backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic==2.10.0
pydantic-settings==2.7.0
google-generativeai==0.8.0
python-dotenv
fastapi-cors
